refactor(webui): PumpSectionの純表示ロジックをdetailLogic.tsへ集約しレビュー指摘を解消
- pumpSectionDisplay/PumpSectionDisplay型をdetailLogic.ts正本へ移し、PumpSection.test.tsをdetailLogic.test.tsへ統合（前例machineStateDisplayと同居、テスト専用exportを解消）
- pumpSection.module.cssの--slot-sizeフォールバックを実トークン値3remへ揃え、デザインwhitelistテストのstyles登録漏れを解消
- blockRegistryCoverage.spec.tsのintentionalGenericをアルファベット順へ整列
