+++ moorestech_web/webui/src/features/blockInventory/details/PumpSection.tsx
@@ -1,20 +1,11 @@
 import { Group, Stack, Text } from "@mantine/core";
-import type { BlockInventoryOpen, PumpDetailData } from "@/bridge";
+import type { BlockInventoryOpen } from "@/bridge";
 import { FluidIcon } from "@/shared/ui";
 import LackHighlightText from "./LackHighlightText";
 import PowerRateText from "./PowerRateText";
-import { machineStateDisplay } from "./detailLogic";
+import { machineStateDisplay, pumpSectionDisplay } from "./detailLogic";
 import { L, useI18n } from "@/shared/i18n";
 import styles from "./pumpSection.module.css";
-
-export type PumpSectionDisplay = { showNoVein: boolean; showPumpingFluids: boolean };
-
-// 汲み上げ対象の有無だけで警告行と流体行を排他に出し分ける（ADR 0051）
-// Whether the pump has targets alone decides between the warning row and the fluid rows (ADR 0051)
-export function pumpSectionDisplay(pump: Pick<PumpDetailData, "pumpingFluids">): PumpSectionDisplay {
-  const hasTargets = pump.pumpingFluids.length > 0;
-  return { showNoVein: !hasTargets, showPumpingFluids: hasTargets };
-}
 
 // ポンプ: 動力行（油井のみ。歯車ポンプは GearSection が担う）+ 公称生成速度 + 鉱脈警告（MinerSection 準拠）
 // Pump: power row (electric pump only; GearSection covers the gear pump), nominal rates, and the vein warning (mirrors MinerSection)
