+++ moorestech_web/webui/src/features/blockInventory/details/detailLogic.test.ts
@@ -6,6 +6,7 @@ import {
   fuelRatio,
   itemsPerMinute,
   machineStateDisplay,
+  pumpSectionDisplay,
   stopReasonTranslationKey,
 } from "./detailLogic";
 
@@ -59,6 +60,22 @@ describe("detailLogic", () => {
     it("要求電力0でも稼働状態なら充足率を表示する", () => {
       expect(machineStateDisplay("idle").showPowerRate).toBe(true);
       expect(machineStateDisplay("processing").showPowerRate).toBe(true);
+    });
+  });
+
+  // 表示分岐は純関数に切り出し、警告行の出し分けをDOM無しで固定する（ADR 0051）
+  // The display branch is a pure function so the warning toggle is pinned without a DOM (ADR 0051)
+  describe("pumpSectionDisplay", () => {
+    it("汲み上げ中流体が空なら警告行だけを出す", () => {
+      const display = pumpSectionDisplay({ pumpingFluids: [] });
+      expect(display.showNoVein).toBe(true);
+      expect(display.showPumpingFluids).toBe(false);
+    });
+
+    it("汲み上げ中流体があれば流体行を出し警告行は出さない", () => {
+      const display = pumpSectionDisplay({ pumpingFluids: [{ fluidId: 1, fluidGuid: "54000000-0000-4000-8000-000000000001", amountPerMinute: 150 }] });
+      expect(display.showNoVein).toBe(false);
+      expect(display.showPumpingFluids).toBe(true);
     });
   });
 });
