+++ moorestech_web/webui/src/features/blockInventory/details/pumpSection.module.css
@@ -1,4 +1,4 @@
 .icon {
-  width: var(--slot-size, 48px);
-  height: var(--slot-size, 48px);
+  width: var(--slot-size, 3rem);
+  height: var(--slot-size, 3rem);
 }
